Flask_Code/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0     
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5       
numpy==1.25.2
praat-parselmouth==0.4.3
scipy==1.14.0
threadpoolctl==3.5.0
Werkzeug==3.0.3
